requirements.txt
# Python dependencies for cloud deployment
ultralytics==8.0.196
flask==3.0.0
pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
torch==2.1.0+cpu
torchvision==0.16.0+cpu
gunicorn==21.2.0
werkzeug==3.0.1

# Find CPU-only PyTorch at:
# https://download.pytorch.org/whl/cpu